Rebuild with Rollup and monolithic package format
Also includes PriorityQueue and Queue utils. This sets up the Rollup
build for a large, dependency-free monolithic package where individual
files can be included as well as the main require or module import
format. With this structure in place, we can start porting the graph
library over from the Ruby prototype.